[WIP] Build from source for OS >= CentOS 6.8 without CUDA (for now).
.. if using Cuda/cudNN I believe we can't target CentOS < 7 which is a
problem for us. There are symbol/ABI hacks detailed on SO which may be
applicable: http://stackoverflow.com/a/34900471, irrespective of cudNN
this recipe generates binaries that cannot be run on CentOS 5.11:

We have built bazel with Conda's gcc-4.8 and the bazel shared libraries
link to that libstdc++. This is in-compatible with the system libstdc++
that Java will load by default, so when the JNI tries to load bazel .so
files, we get missing symbols. To avoid this LD_LIBRARY_PATH is used to
force Java to load Conda's libstdc++.so from our libgcc package(s).

The proper fix is to add an openjdk-8 package and build bazel with that.
They would both be set to depend on Conda's gcc-4.8 and libgcc packages.

As things stand, because openjdk-8 is not available on CentOS 5.11, the
bazel build was done on CentOS 6.8. This means the CentOS 6.8 openjdk-8
glibc 2.9 minimum version requirement gets passed on to bazel and later
that is passed on by bazel's swig mechanism into _pywrap_tensorflow.so.

Note: I have added swig as a dependency here, but am not sure that it's
used. This is something I will check into, hence marking this commit as
WIP.